// src/components/NewsSnippet/interfaces.ts
// интерфейсы для типов данных

export interface IData_TagItem {
  value: string; // название тега
  count: number; // кол-во тегов с указанным названием
}

export interface IData_TrafficItem {
  value: string; // название страны-источник траффика
  count: number; // объём траффика для указанной страны доля/процент
}

export interface IData_SnippetNews {
  ID: number; // идентификатор новости
  TI: string; // заголовок новости
  AB: string; // содержимое новости (полное)
  URL: string; // ссылка на новость
  DOM: string; // домен
  DP: string; // дата и время публикации новости в формате "%Y-%m-%dT%H:%M:%S"
  LANG: string; // язык новости
  REACH: number; // охват новости
  KW: IData_TagItem[]; // ключевые слова
  AU: string[]; // автор новости
  CNTR: string; // страна
  CNTR_CODE: string; // код страны
  SENT: string; // сантимент новости (например, 'positive', 'negative', 'neutral')
  TRAFFIC: IData_TrafficItem[]; // траффик из стран
  FAV: string; // ссылка на иконку (может быть относительной)
  HIGHLIGHTS: string[]; // блоки содержимого новости с ключевыми словами <kw>...</kw>
}